[IMP] account: stronger constraint on repartition lines
We now force tax repartition lines to use the same percentages for invoices and credit notes. Thanks to that, we are able to simplify the way manual modifications made on vendor bills' account.invoice.tax entries must affect the ones of the refunds created for them. This also prepares the code for the incoming big accounting revamping.

closes odoo/odoo#33577

Signed-off-by: Laurent Smet <smetl@users.noreply.github.com>